lab9/python/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
grpcio==1.62.1
grpcio-tools==1.62.1
sympy==1.12